backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pillow==10.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
